requirements.txt
discord.py>=2.4.0
iracingdataapi>=2.0.0
python-dotenv>=1.0.0
aiosqlite>=0.20.0